+++ src/lib.rs
@@ -1,0 +1,8 @@
+#![feature(let_else)]
+#![allow(clippy::if_not_else)]
+
+pub mod bridge;
+pub mod channel;
+pub mod crdt_client;
+pub mod ot_client;
+pub mod ot_server;
